requirements.txt
streamlit
openai-whisper
moviepy
googletrans
pandas
sqlite3
